requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
autopep8==2.0.4
billiard==4.2.0
blessed==1.20.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0
django-picklefield==3.1
django-q==1.3.9
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.2
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.19.0
jedi==0.19.1
kombu==5.3.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
